Extract saveWorkState()
